Relocate concrete DeleteLogic + IDeleteLogic into headless Gum.Presentation (ADR-0005 Phase 3)
Continues the DeleteLogic decoupling from #3225's last progress update: narrow ports (IDeletePluginNotifier, IDeleteProjectProvider), the IReferenceFinder relocation, and the IDeleteDialogService seam already left DeleteLogic WPF-free; this PR does the git mv.

Two dead extension-method dependencies blocked a clean move and are seam-sealed here rather than deferred:
- IsScreensFolderTreeNode/IsComponentsFolderTreeNode/IsTopScreenContainerTreeNode/IsTopComponentContainerTreeNode: the ITreeNode-typed overloads in ElementTreeViewManager.cs only unwrapped to the WinForms TreeNode overload. Reimplemented directly against ITreeNode.Parent/.Tag/.Text (TreeNodeWrapper.Parent already wraps Node.Parent recursively, so results are identical for live tool nodes, and now also correct for any other ITreeNode implementation instead of always returning false) and relocated into Gum.Presentation. The WinForms TreeNode overloads stay put.
- IDialogService.ShowYesNoMessage: pure wrapper over ShowMessage with zero WPF dependency; split out of the WPF-project IDialogServiceExt into its own headless extension class, leaving Show<T>/ShowChoices<T> behind pending their dialog-viewmodel dependencies going headless.

Also dropped a dead 'using Gum.Gui.Windows' from IDeleteLogic.cs and dead Gum.Plugins/Gum.Services/Gum.Undo usings from DeleteLogic.cs.

Mirrors the UndoManager relocation (#3357): granted Gum.Presentation.csproj InternalsVisibleTo(GumToolUnitTests) so DeleteLogicTests keeps reaching DeleteLogic's internal members unchanged; drop that grant once the tests migrate into Gum.Presentation.Tests (mirroring UndoManagerTests' #3358).

Behavior-preserving. Verified: Gum.Presentation compiles standalone (0 errors); GumFull.sln builds (0 errors); GumToolUnitTests 1067 passed / 0 failed / 4 skipped; Gum.Presentation.Tests 42 passed / 0 failed.

Part of #3225 (ADR-0005 Phase 3 headless-assembly extraction).
