+++ Tools/Gum.Presentation/Managers/DeleteLogic.cs
@@ -3,11 +3,8 @@ using Gum.DataTypes;
 using Gum.DataTypes.Behaviors;
 using Gum.DataTypes.Variables;
 using Gum.Logic;
-using Gum.Plugins;
-using Gum.Services;
 using Gum.Services.Dialogs;
 using Gum.ToolStates;
-using Gum.Undo;
 using Gum.Wireframe;
 using System;
 using System.Collections;
