+++ Tools/Gum.Presentation/Managers/IDeleteLogic.cs
@@ -1,7 +1,6 @@
 using Gum.DataTypes;
 using Gum.DataTypes.Behaviors;
 using Gum.DataTypes.Variables;
-using Gum.Gui.Windows;
 using System.Collections.Generic;
 
 namespace Gum.Managers;
